Code cleanup and get LEDs working
