+++ src/Manchester_State.c
@@ -6,6 +6,7 @@
  */
 
 #include "Manchester_State.h"
+#include "gpio.h"
 
 static uint8_t RX;
 static volatile EXTI *EXTI0 = (EXTI *) 0x40013C00;
@@ -13,9 +14,15 @@ static volatile GPIOx *GPIOC = (GPIOx *) 0x40020800;
 
 static volatile enum STATES state;
 
+static void initLEDs();
+static void setLED(enum STATES state);
+
 void init_state(){
 	state = IDLE;
 	RX = 1;
+	initLEDs();
+
+	setLED(state);
 
 	//start channel interrupt and timer interrupt
 	init_RX_channel();
@@ -24,6 +31,33 @@ void init_state(){
 
 enum STATES getState(){
 	return state;
+}
+
+static void initLEDs(){
+	//enable GPIOC clock
+	enable_clock('C');
+
+	//GPIOC clock already selected
+	//set GPIOC pin 1,2,3 to output for LEDs
+	set_pin_mode('C', 1, OUTPUT);
+	set_pin_mode('C', 2, OUTPUT);
+	set_pin_mode('C', 3, OUTPUT);
+}
+
+static void setLED(enum STATES state){
+	switch(state){
+		case IDLE:
+			GPIOC -> BSSR |= (((0b110 << 1) << 16) | (0b001 << 1));
+			break;
+		case BUSY:
+			GPIOC -> BSSR |= (((0b101 << 1) << 16) | (0b010 << 1));
+			break;
+		case COLLISION:
+			GPIOC -> BSSR |= (((0b011 << 1) << 16) | (0b100 << 1));
+			break;
+		default:
+			break;
+	}
 }
 
 //Timer has expired
@@ -35,7 +69,11 @@ void TIM2_IRQHandler(void){
 		state = IDLE;
 	}
 
+	// Obtain the most recent RX value
 	RX = (GPIOC -> IDR) & (1 << 0);
+
+	// Set LEDs to match state
+	setLED(state);
 
 	//clear flag
 	*(TIM2_SR) &= ~(1<<2);
@@ -51,6 +89,9 @@ void EXTI0_IRQHandler(void){
 
 	//capture RX
 	RX = (GPIOC -> IDR) & (1 << 0);
+
+	// Set LEDs to match state
+	setLED(state);
 
 	//clear the status bits for the timer and RX
 	*(TIM2_SR) &= ~(1<<2);
